==> code/02-analyses.Rmd <==
---
title: "Analyses"
author: "Stefano Coretta"
date: "10/09/2021"
output:
  html_document:
    number_sections: true
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
knitr::opts_knit$set(root.dir = here::here())
library(tidyverse)
theme_set(theme_light())
library(magrittr)
library(FactoMineR)
library(factoextra)
library(brms)
library(rnaturalearth)
library(sf)
library(tidygeocoder)
library(HH)
```

# Read data

```{r}
emilianto <- readRDS("data/rds/emilianto.rds")
emilianto_attitude <- readRDS("data/rds/emilianto_attitude.rds")
attitudes <- readRDS("data/rds/attitudes.rds")
pa_spaces <- readRDS("data/rds/pa_spaces.rds")
ty_spaces <- readRDS("data/rds/ty_spaces.rds")
```


# Emilian

```{r emilian}
emilian <- emilianto %>%
  filter(language == "Emilian")
```

## Participants

```{r em-gender}
emilian %>%
  ggplot(aes(gender)) +
  geom_bar()
```

```{r em-age}
emilian %>%
  ggplot(aes(age)) +
  geom_histogram()

emilian %>%
  ggplot(aes(age)) +
  geom_density()
```

```{r em-education}
emilian %>%
  ggplot(aes(education)) +
  geom_bar()
```

```{r em-age-edu}
emilian %>%
  ggplot(aes(age, education)) +
  geom_jitter(height = 0.2, alpha = 0.5)
```

```{r em-occupation}
emilian %>%
  count(profession) %>%
  ggplot(aes(reorder(profession, -n), n)) +
  geom_bar(stat = "identity")
```

```{r em-languages-family}
emilian %>%
  count(languages_family) %>%
  ggplot(aes(reorder(languages_family, -n), n)) +
  geom_bar(stat = "identity")
```

```{r em-languages-parents}
emilian %>%
  count(languages_parents) %>%
  ggplot(aes(reorder(languages_parents, -n), n)) +
  geom_bar(stat = "identity")
```

## Language

### Understand

```{r em-understand}
emilian %>%
  ggplot(aes(understand, fill = understand)) +
  geom_bar() +
  scale_fill_brewer(type = "div") +
  theme_dark()
```

```{r em-understand-gender}
emilian %>%
  ggplot(aes(understand, fill = gender)) +
  geom_bar()

emilian %>%
  ggplot(aes(understand, fill = gender)) +
  geom_bar(position = "fill")
```

```{r em-understand-age}
emilian %>%
  ggplot(aes(age, fill = understand)) +
  geom_histogram(binwidth = 5) +
  facet_grid(understand ~ .)
```

```{r em-understand-profession}
emilian %>%
  ggplot(aes(understand, fill = profession)) +
  geom_bar()

emilian %>%
  ggplot(aes(understand, fill = profession)) +
  geom_bar(position = "fill")
```

### Speak

```{r em-speak}
emilian %>%
  ggplot(aes(speak, fill = speak)) +
  geom_bar() +
  scale_fill_brewer(type = "div")
```

```{r em-speak-gender}
emilian %>%
  ggplot(aes(speak, fill = gender)) +
  geom_bar()

emilian %>%
  ggplot(aes(speak, fill = gender)) +
  geom_bar(position = "fill")
```

```{r em-speak-age}
emilian %>%
  ggplot(aes(age, fill = speak)) +
  geom_histogram(binwidth = 5) +
  facet_grid(speak ~ .)
```

```{r em-speak-profession}
emilian %>%
  ggplot(aes(speak, fill = profession)) +
  geom_bar()

emilian %>%
  ggplot(aes(speak, fill = profession)) +
  geom_bar(position = "fill")
```



### Read and write

```{r em-read-write}
emilian %>%
  ggplot(aes(read_write, fill = read_write)) +
  geom_bar()
```

```{r em-read-write-gender}
emilian %>%
  ggplot(aes(read_write, fill = gender)) +
  geom_bar()

emilian %>%
  ggplot(aes(read_write, fill = gender)) +
  geom_bar(position = "fill")
```

```{r em-read-write-age}
emilian %>%
  drop_na(read_write) %>% 
  ggplot(aes(age, fill = read_write)) +
  geom_histogram(binwidth = 5) +
  facet_grid(read_write ~ .)
```

```{r em-read-write-profession}
emilian %>%
  ggplot(aes(read_write, fill = profession)) +
  geom_bar()

emilian %>%
  ggplot(aes(read_write, fill = profession)) +
  geom_bar(position = "fill")
```




### Attitude

```{r em-attitude}
emilian %>%
  select(educated:familiar) %>%
  pivot_longer(educated:familiar, names_to = "feature", values_to = "rating") %>%
  ggplot(aes(as.factor(rating), fill = as.factor(rating))) +
  geom_bar() +
  scale_fill_brewer() +
  facet_grid(. ~ feature)

emilian %>%
  select(educated:familiar) %>%
  pivot_longer(educated:familiar, names_to = "feature", values_to = "rating") %>%
  ggplot(aes(feature, fill = as.factor(rating))) +
  geom_bar(position = "fill") +
  scale_fill_brewer()
```

## Urban vs Rural

```{r em-rur}
emil_rur <- emilian %>%
  mutate(
    ru_ur = ifelse(
      str_detect(birth_place, "-RU"), "rural",
      ifelse(
        str_detect(birth_place, "-UR"), "urban",
        NA
      )
    )
  )

emil_rur_clean <- emil_rur %>%
  select(id, understand, speak, read_write, educated:familiar, ru_ur) %>%
  mutate(
    understand = ordered(understand, levels = c("NO", "AL", "50/50", "G", "VG")),
    speak = ordered(speak, levels = c("NO", "AL", "50/50", "G", "VG")),
    across(educated:familiar, ~ as.ordered(.x))
  ) %>%
 drop_na()
```

```{r em-rur-plot}
emil_rur %>%
  drop_na(ru_ur) %>%
  ggplot(aes(ru_ur, fill = understand)) +
  geom_bar()

emil_rur %>%
  drop_na(ru_ur) %>%
  ggplot(aes(ru_ur, fill = understand)) +
  geom_bar(position = "fill")
```


# Esperanto

```{r esperanto}
esperanto <- emilianto %>%
  filter(language == "Esperanto")
```

## Participants

```{r eo-gender}
esperanto %>%
  ggplot(aes(gender)) +
  geom_bar()
```

```{r eo-age}
esperanto %>%
  ggplot(aes(age)) +
  geom_histogram()

esperanto %>%
  ggplot(aes(age)) +
  geom_density()
```

```{r eo-education}
esperanto %>%
  ggplot(aes(education)) +
  geom_bar()
```

```{r eo-age-edu}
esperanto %>%
  ggplot(aes(age, education)) +
  geom_jitter(height = 0.2, alpha = 0.5)
```

```{r eo-occupation}
esperanto %>%
  count(profession) %>%
  ggplot(aes(reorder(profession, -n), n)) +
  geom_bar(stat = "identity")
```

```{r eo-languages-family}
esperanto %>%
  count(languages_family) %>%
  ggplot(aes(reorder(languages_family, -n), n)) +
  geom_bar(stat = "identity")
```


## Language

### Understand

```{r eo-understand}
esperanto %>%
  ggplot(aes(understand, fill = understand)) +
  geom_bar() +
  scale_fill_brewer(type = "div") +
  theme_dark()
```

```{r eo-understand-gender}
esperanto %>%
  ggplot(aes(understand, fill = gender)) +
  geom_bar()

esperanto %>%
  ggplot(aes(understand, fill = gender)) +
  geom_bar(position = "fill")
```

```{r eo-understand-age}
esperanto %>%
  ggplot(aes(age, fill = understand)) +
  geom_histogram(binwidth = 5) +
  facet_grid(understand ~ .)
```

```{r eo-understand-profession}
esperanto %>%
  ggplot(aes(understand, fill = profession)) +
  geom_bar()

esperanto %>%
  ggplot(aes(understand, fill = profession)) +
  geom_bar(position = "fill")
```


### Speak

```{r eo-speak}
esperanto %>%
  ggplot(aes(speak, fill = speak)) +
  geom_bar() +
  scale_fill_brewer(type = "div")
```


```{r eo-speak-gender}
esperanto %>%
  ggplot(aes(speak, fill = gender)) +
  geom_bar()

esperanto %>%
  ggplot(aes(speak, fill = gender)) +
  geom_bar(position = "fill")
```

```{r eo-speak-age}
esperanto %>%
  ggplot(aes(age, fill = speak)) +
  geom_histogram(binwidth = 5) +
  facet_grid(speak ~ .)
```

```{r eo-speak-profession}
esperanto %>%
  ggplot(aes(speak, fill = profession)) +
  geom_bar()

esperanto %>%
  ggplot(aes(speak, fill = profession)) +
  geom_bar(position = "fill")
```

### Read and write

```{r eo-read-write}
esperanto %>%
  ggplot(aes(read_write, fill = read_write)) +
  geom_bar()
```


### Attitude

```{r eo-attitude}
esperanto %>%
  select(educated:familiar) %>%
  pivot_longer(educated:familiar, names_to = "feature", values_to = "rating") %>%
  drop_na() %>%
  ggplot(aes(as.factor(rating), fill = as.factor(rating))) +
  geom_bar() +
  scale_fill_brewer() +
  facet_grid(. ~ feature)

esperanto %>%
  select(educated:familiar) %>%
  pivot_longer(educated:familiar, names_to = "feature", values_to = "rating") %>%
  drop_na() %>%
  ggplot(aes(feature, fill = as.factor(rating))) +
  geom_bar(position = "fill") +
  scale_fill_brewer()
```

# Dim-1 and age

```{r dim1-age}
emilianto_attitude %>%
  ggplot(aes(age, dim_1, colour = language)) +
  geom_point() +
  geom_smooth(method = "lm", formula = "y ~ x")
```

# Locations

```{r geo}
if (file.exists("./data/raw/geo.csv")) {
  cat("Reading geocoding...\n")
  geo <- read_csv("./data/raw/geo.csv")
} else {
  birth_em <- emilian %>% select(birth_place_it) %>% unique()
  geo <- geocode(birth_em, city = birth_place_it, method = "osm", verbose = TRUE)
  write_csv(geo, file = "./data/raw/geo.csv")
}
```


```{r europe}
europe <- ne_countries(continent = "Europe", returnclass = "sf", scale = "medium")
```

```{r emilian-map}
ggplot() +
  geom_sf(data = europe) +
  geom_point(data = geo, aes(long, lat)) +
  coord_sf(xlim = c(7, 14), ylim = c(43, 47))
```


# Modelling

## Understand

```{r understand-bm-prior}
get_prior(
  understand ~
    language * dim_1,
  data = emilianto_attitude,
  family = cumulative()
)

understand_priors <- c(
  prior(normal(0, 3), class = Intercept),
  prior(normal(0, 1), class = b)
)

understand_pchek <- brm(
  understand ~
    language * dim_1,
  data = emilianto_attitude,
  family = cumulative(),
  prior = understand_priors,
  sample_prior = "only",
  backend = "cmdstanr",
  cores = 4,
  file = "./data/rds/understand_pchek"
)

conditional_effects(understand_pchek, effects = "dim_1", conditions = make_conditions(emilianto_attitude, "language"), categorical = TRUE)
```


```{r understand-bm}
understand_bm <- brm(
  understand ~
    language * dim_1,
  data = emilianto_attitude,
  family = cumulative(),
  prior = understand_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/rds/understand_bm"
)
```

```{r understand-bm-sum}
understand_bm
summary(understand_bm, prob = 0.7)
```


```{r understand-bm-cond}
ubc <- conditional_effects(understand_bm, effects = "dim_1", categorical = TRUE, conditions = make_conditions(emilianto_attitude, "language"), plot = FALSE)
```

```{r understand-bm-cond-plot}
plot(ubc, plot = FALSE)[[1]] +
  geom_vline(xintercept = 0, linetype = "dashed") +
  scale_fill_brewer(type = "div", palette = "PRGn") +
  scale_color_brewer(type = "div", palette = "PRGn") +
  labs(
    x = "Attitude (MCA Dimension 1)",
    y = "Probability",
    caption = "Attidude is operationalised as the scores of Dimension 1 from the Multiple Correspondence Analysis.\nGreater values correspond to a more positive attitude.",
    colour = "Understand\ncompetence",
    fill = "Understand\ncompetence"
  )

ggsave("img/understand-brm.png", width = 7, height = 5)
```


## Speak

```{r speak-bm}
speak_bm <- brm(
  speak ~
    language * dim_1,
  data = emilianto_attitude,
  family = cumulative(),
  # Same priors as understand_bm
  prior = understand_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/rds/speak_bm"
)
```

```{r speak-bm-sum}
speak_bm
summary(speak_bm, prob = 0.4)
```


```{r speak-bm-cond}
sbc <- conditional_effects(speak_bm, effects = "dim_1", categorical = TRUE, conditions = make_conditions(speak_bm, "language"))
```

```{r speak-bm-cond-plot}
plot(sbc, plot = FALSE)[[1]] +
  scale_fill_brewer(type = "div", palette = "PRGn") +
  scale_color_brewer(type = "div", palette = "PRGn") +
  labs(
    x = "Attitude (MCA Dimension 1)",
    y = "Probability",
    caption = "Attidude is operationalised as the scores of Dimension 1 from the Multiple Correspondence Analysis.\nGreater values correspond to a more positive attitude.",
    colour = "Speaking\ncompetence",
    fill = "Speaking\ncompetence"
  )

ggsave("img/speak-brm.png", width = 7, height = 5)
```


## Read and write

```{r rw-bm}
rw_bm <- brm(
  read_write ~
    dim_1,
  data = emilianto_attitude %>% filter(language == "Emilian"),
  family = bernoulli(),
  # Same priors as understand_bm
  prior = understand_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/rds/rw_bm"
)
```

```{r rw-bm-sum}
rw_bm
```


```{r rw-bm-cond}
conditional_effects(rw_bm, effects = "dim_1")
```

## Spaces

```{r spaces-bm}
a_spaces <- pa_spaces %>%
  filter(spaces == "active") %>%
  mutate(
    speak_bi = ifelse(speak %in% c("V", "VG"), "high", "low"),
    speak_bi = factor(speak_bi, levels = c("low", "high")),
    proportion_10 = proportion * 10
  )

spaces_bm <- brm(
  speak_bi ~
    language * proportion_10,
  data = a_spaces,
  family = bernoulli(),
  # Same priors as understand_bm
  prior = understand_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/rds/spaces_bm"
)
```

```{r spaces-bm-sum}
spaces_bm
```

```{r spaces-bm-cond}
spbc <- conditional_effects(spaces_bm, "proportion_10:language")
```

```{r spaces-bm-cond-plot}
plot(spbc, plot = FALSE)[[1]] +
  scale_y_continuous(limits = c(0, 1)) +
  scale_colour_brewer(type = "qual") +
  scale_fill_brewer(type = "qual") +
  labs(
    x = "Spaces (score)",
    y = "Speaking competence (low vs high)",
    caption = str_wrap("Speaking competence was binned (low = NO/AL/50:50, high = G/VG.")
  )

ggsave("img/spaces-brm.png", width = 7, height = 5)
```


## Mean competence

```{r meanc}
emilianto_attitude %>%
  ggplot(aes(mean_comp)) +
  geom_density() +
  geom_rug()
```

```{r meanc-jitter}
emilianto_attitude %>%
  ggplot(aes(language, mean_comp)) +
  geom_jitter(width = 0.2, alpha = 0.5, height = 0) +
  expand_limits(y = 0)
```   

```{r mean-jitter-dim}
emilianto_attitude %>%
  ggplot(aes(language, mean_comp, colour = dim_1)) +
  geom_jitter(width = 0.2, alpha = 0.5, height = 0) +
  expand_limits(y = 0)
```

```{r meanc-dim1}
emilianto_attitude %>%
  ggplot(aes(dim_1, mean_comp)) +
  geom_point() +
  facet_grid(~ language)
```

```{r meanc-bm}
mean_comp_bm <- brm(
  mean_comp ~
    language *
    dim_1,
  data = emilianto_attitude,
  family = zero_one_inflated_beta(),
  backend = "cmdstanr",
  cores = 4,
  file = "./data/rds/mean_comp_bm"
)

mean_comp_bm
```

```{r meanc-bm-cond}
conditional_effects(mean_comp_bm, effects = "dim_1:language")
```


## Rural vs urban

### Understand

```{r understand-rur}
understand_rur <- emil_rur_clean %>%
  count(understand, ru_ur) %>%
  group_by(ru_ur) %>%
  mutate(
    ru_ur_prop = n / sum(n)
  ) %>%
  ungroup() %>%
  mutate(
    understand = factor(understand, ordered = FALSE),
    ru_ur = factor(ru_ur, levels = c("rural", "urban"))
  )
```

```{r understand-rur-priors}
get_prior(
  n ~
    understand * ru_ur,
  data = understand_rur,
  family = Beta()
)
```

```{r understand-rur-bm}
understand_rur_priors <- c(
  prior(normal(0, 3), class = Intercept),
  prior(normal(0, 1), class = b)
)

understand_rur_bm <- brm(
  ru_ur_prop ~
    understand * ru_ur,
  data = understand_rur,
  family = Beta(),
  prior = understand_rur_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/rds/understand_rur_bm"
)

understand_rur_bm
```

```{r understand-rur-cond}
conditional_effects(understand_rur_bm, "understand:ru_ur")
```

### Speak

```{r speak-rur}
speak_rur <- emil_rur_clean %>%
  count(speak, ru_ur) %>%
  group_by(ru_ur) %>%
  mutate(
    ru_ur_prop = n / sum(n)
  ) %>%
  ungroup() %>%
  mutate(
    speak = factor(speak, ordered = FALSE),
    ru_ur = factor(ru_ur, levels = c("rural", "urban"))
  )
```


```{r speak-rur-bm}
speak_rur_bm <- brm(
  ru_ur_prop ~
    speak * ru_ur,
  data = speak_rur,
  family = Beta(),
  prior = understand_rur_priors,
  backend = "cmdstanr",
  cores = 4,
  file = "./data/rds/speak_rur_bm"
)

speak_rur_bm
```

```{r speak-rur-cond}
conditional_effects(speak_rur_bm, "speak:ru_ur")
```

### Read and write

```{r rw-rur}
rw_rur <- emil_rur_clean %>%
  count(read_write, ru_ur) %>%
  group_by(ru_ur) %>%
  mutate(
    ru_ur_prop = n / sum(n)
  ) %>%
  ungroup() %>%
  mutate(
    ru_ur = factor(ru_ur, levels = c("rural", "urban"))
  )
```


```{r rw-rur-bm}
rw_rur_bm <- brm(
  ru_ur_prop ~
    read_write * ru_ur,
  data = rw_rur,
  family = Beta(),
  prior = understand_rur_priors,
  backend = "cmdstanr",
  cores = 4,
  file = "./data/rds/rw_rur_bm"
)

rw_rur_bm
```

```{r rw-rur-cond}
conditional_effects(rw_rur_bm, "read_write:ru_ur")
```

# Diverging stacked bar chart

```{r div-stacked}
emilianto_lik <- emilianto_attitude %>%
  dplyr::select(language, educated:familiar) %>%
  pivot_longer(educated:familiar, names_to = "quality", values_to = "score") %>%
  count(language, quality, score) %>%
  pivot_wider(names_from = "score", values_from = n)

likert(
  quality ~ . | language,
  emilianto_lik,
  as.percent = TRUE,
  main = "Diverging staked bar chart"
)
```

# Tables and plots included in the paper

```{r tables}
columns <- c("gender", "age_2", "profession_2", "languages_family_2")

for (column in columns) {
  emilianto %>%
    group_by(language) %>%
    count(.data[[column]]) %>%
    mutate(
      prop = round(n / sum(n, na.rm = TRUE) * 100, 1)
    ) %>%
    pivot_wider(names_from = .data[[column]], values_from = c(n, prop)) %>%
    print()
}
```


```{r tables-2}
columns <- c("gender", "age_2", "profession_2", "languages_family_2")

for (column in columns) {
  emilianto %>%
    filter(
      understand %in% c("50/50", "G", "VG"),
      speak %in% c("50/50", "G", "VG")
    ) %>%
    group_by(language) %>%
    count(.data[[column]]) %>%
    mutate(
      prop = round(n / sum(n, na.rm = TRUE) * 100, 1)
    ) %>%
    pivot_wider(names_from = .data[[column]], values_from = c(n, prop)) %>%
    print()
}
```

```{r tables-3}
columns <- c("gender", "age_2", "profession_2", "languages_family_2")

emilianto_50 <- emilianto %>%
  mutate(
    und_speak = case_when(
      understand %in% c("50/50", "G", "VG") ~ ">50",
      speak %in% c("50/50", "G", "VG") ~ ">50",
      TRUE ~ "<50"
    )
  )

for (column in columns) {
 emilianto_50 %>%
    group_by(language) %>%
    add_count(name = "tot") %>%
    ungroup() %>%
    count(.data[[column]], und_speak, language, tot) %>%
    mutate(
      prop = round((n / tot) * 100, 1)
    ) %>%
    pivot_wider(names_from = .data[[column]], values_from = c(n, prop)) %>%
    print()
}
```

```{r pap-plot}
pa_spaces %>%
  ggplot(aes(gender, proportion, fill = gender)) +
  geom_hline(yintercept = 0, size = 0.25) +
  geom_hline(yintercept = 1, size = 0.25) +
  geom_violin(colour = NA, bw = 0.1) +
  geom_boxplot(width = 0.05, fill = "black", colour = "black") +
  stat_summary(colour = "white", fun = "median", geom = "point") +
  ylim(0, 1) +
  facet_grid(spaces ~ language, labeller = as_labeller(list("competence" = c("a", "b")))) +
  scale_fill_brewer(type = "qual", palette = "Dark2") +
  labs(
    y = "Spaces (proportion)",
    x = "Gender"
  ) +
  theme_minimal(base_size = 16) +
  theme(legend.position = "none")

ggsave("./img/spaces.png", width = 7, height = 5)
```

```{r ty-plot}
ty_spaces %>%
  ggplot(aes(types, proportion, fill = types)) +
  geom_hline(yintercept = 0, size = 0.25) +
  geom_hline(yintercept = 1, size = 0.25) +
  geom_violin(colour = NA) +
  geom_boxplot(width = 0.05, fill = "black", colour = "black") +
  stat_summary(colour = "white", fun = "median", geom = "point") +
  ylim(0, 1) +
  facet_grid(~ language) +
  scale_fill_brewer(type = "qual", palette = "Dark2") +
  labs(
    y = "Total spaces (proportion)",
    x = "Space type"
  ) +
  theme_minimal(base_size = 16) +
  theme(legend.position = "none")

ggsave("./img/space-types.png", width = 7, height = 5)
```

```{r ty-perc}
ty_spaces %>%
  dplyr::select(language, id) %>%
  distinct() %>%
  count(language)

n_spaces <- c(42, 2, 2, 13, 23, 12, 11)

n_spaces_em <- n_spaces * 434
n_spaces_eo <- n_spaces * 150

ty_spaces_count <- ty_spaces %>%
  dplyr::select(language, id, ends_with("_spaces")) %>%
  distinct() %>%
  pivot_longer(passive_spaces:other_spaces, names_to = "space_type", values_to = "count") %>%
  group_by(language, space_type) %>%
  summarise(
    total_count = sum(count),
    .groups = "drop"
  )

ty_spaces_count$n_spaces <- c(n_spaces_em, n_spaces_eo)

ty_spaces_perc <- ty_spaces_count %>%
  mutate(
    perc = round((total_count / n_spaces) * 100)
  ) %>%
  mutate(space_type = str_remove(space_type, "_spaces"))
```

```{r ty-spaces-plot}
ty_spaces_perc %>%
  filter(!(space_type %in% c("passive", "active"))) %>%
  mutate(
    space_type = factor(space_type, levels = c("physical", "media", "press", "virtual", "other"))
  ) %>%
  ggplot(aes(space_type, perc, fill = space_type)) +
  geom_bar(stat = "identity") +
  scale_fill_brewer(type = "qual", palette = "Dark2") +
  facet_grid(~ language) +
  ylim(0, 100) +
  labs(x = "Space type", y = "Percent") +
  theme(legend.position = "none")

ggsave("./img/space-types-perc.png", width = 7, height = 5)
```
