==> code/03-modelling.Rmd <==
---
title: "Analyses"
author: "Stefano Coretta"
date: "10/09/2021"
output:
  html_document:
    number_sections: true
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
knitr::opts_knit$set(root.dir = here::here())
library(tidyverse)
theme_set(theme_light())
library(magrittr)
library(FactoMineR)
library(factoextra)
library(brms)
library(rnaturalearth)
library(sf)
library(tidygeocoder)
library(HH)
library(usdm)
```

# Read data

```{r}
emilianto <- readRDS("data/rds/emilianto.rds")
emilianto_attitude <- readRDS("data/rds/emilianto_attitude.rds") %>%
  mutate(
    language = factor(language)
  )
contrasts(emilianto_attitude$language) <- "contr.sum"
attitudes <- readRDS("data/rds/attitudes.rds")
pa_spaces <- readRDS("data/rds/pa_spaces.rds")
ty_spaces <- readRDS("data/rds/ty_spaces.rds")
```


# Checks

```{r vif}
eea <- emilianto_attitude %>% dplyr::select(dim_1, passive_prop, active_prop)
vif(as.data.frame(eea))
cor(emilianto_attitude$passive_prop, emilianto_attitude$active_prop, method = "spearman")

eea <- emilianto_attitude %>% dplyr::select(dim_1, spaces_prop)
vif(as.data.frame(eea))
cor(emilianto_attitude$dim_1, emilianto_attitude$spaces_prop, method = "spearman")
```

Since `passive_prop` and `active_prop` have a somewhat high correlation (about 0.7), we decided to merge active and passive spaces into a single measure of spaces proportion `spaces_prop`.

The correlation between `spaces_prop` and `dim_1` is 0.26. The VIF is very close to 1. We believe it is safe to include both variables in the model.

```{r dim-1}
emilianto_attitude %>%
  ggplot(aes(comprehend, dim_1, colour = comprehend)) +
  geom_violin() +
  geom_jitter(width = 0.1, alpha = 0.5) +
  facet_grid(~language)
emilianto_attitude %>%
  ggplot(aes(speak, dim_1, colour = speak)) +
  geom_violin() +
  geom_jitter(width = 0.1, alpha = 0.5) +
  facet_grid(~language)
```


# Modelling

## comprehend

```{r comprehend-bm-prior}
get_prior(
  comprehend ~
    language * dim_1 * spaces_prop,
  data = emilianto_attitude,
  family = cumulative()
)

comprehend_priors <- c(
  prior(normal(0, 3), class = Intercept),
  prior(normal(0, 1), class = b)
)

comprehend_pcheck <- brm(
  comprehend ~
    language * dim_1 * spaces_prop,
  data = emilianto_attitude,
  family = cumulative(),
  prior = comprehend_priors,
  sample_prior = "only",
  backend = "cmdstanr",
  cores = 4,
  file = "./data/cache/comprehend_pcheck"
)

conditional_effects(comprehend_pcheck, "spaces_prop", conditions = make_conditions(emilianto_attitude, "language", "dim_1"), categorical = TRUE)
```


```{r comprehend-bm}
comprehend_bm <- brm(
  comprehend ~
    language * dim_1 * spaces_prop,
  data = emilianto_attitude,
  family = cumulative(),
  prior = comprehend_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/cache/comprehend_bm"
)
```

```{r comprehend-bm-sum}
comprehend_bm
```

```{r ubm-cond-spaces}
conditions <- expand.grid(
  dim_1 = c(-2.5, -1.5, 0, 1.5),
  language = c("Emilian", "Esperanto")
)

rownames(conditions) <- unite(conditions, "cond__", everything()) %>% pull(cond__)

ubm_cond_sp <- conditional_effects(comprehend_bm, effects = "spaces_prop", categorical = TRUE, conditions = conditions)
plot(ubm_cond_sp, facet_args = list(ncol = 4))[[1]] +
  labs(fill = "Comprehension\nlevel", colour =  "Comprehension\nlevel", x = "Dimension 1 (language attitudes)")


ggsave("img/compr-brm-sp.png", width = 7, height = 4)
```

```{r ubm-cond-dim}
conditions <- expand.grid(
  spaces_prop = c(0, 0.25, 0.5, 0.75, 1),
  language = c("Emilian", "Esperanto")
)

rownames(conditions) <- unite(conditions, "cond__", everything()) %>% pull(cond__)

ubm_cond_dim <- conditional_effects(comprehend_bm, effects = "dim_1", categorical = TRUE, conditions = conditions)
plot(ubm_cond_dim, facet_args = list(ncol = 5))[[1]] +
  labs(fill = "Comprehension\nlevel", colour =  "Comprehension\nlevel", x = "Dimension 1 (language attitudes)")

ggsave("img/compr-brm-dim.png", width = 7, height = 4)
```


## Speak

```{r speak-bm}
speak_bm <- brm(
  speak ~
    language * dim_1 * spaces_prop,
  data = emilianto_attitude,
  family = cumulative(),
  # Same priors as comprehend_bm
  prior = comprehend_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/cache/speak_bm"
)
```

```{r speak-bm-sum}
speak_bm
```


```{r sbm-cond-spaces}
conditions <- expand.grid(
  dim_1 = c(-2.5, -1.5, 0, 1.5),
  language = c("Emilian", "Esperanto")
)

rownames(conditions) <- unite(conditions, "cond__", everything()) %>% pull(cond__)

sbm_cond_sp <- conditional_effects(speak_bm, effects = "spaces_prop", categorical = TRUE, conditions = conditions)
plot(sbm_cond_sp, facet_args = list(ncol = 4))[[1]] +
  labs(fill = "Speaking\nlevel", colour = "Speaking\nlevel", x = "Proportion of spaces")

ggsave("img/speak-brm-sp.png", width = 7, height = 4)
```

```{r sbm-cond-dim}
conditions <- expand.grid(
  spaces_prop = c(0, 0.25, 0.5, 0.75, 1),
  language = c("Emilian", "Esperanto")
)

rownames(conditions) <- unite(conditions, "cond__", everything()) %>% pull(cond__)

sbm_cond_dim <- conditional_effects(speak_bm, effects = "dim_1", categorical = TRUE, conditions = conditions)
plot(sbm_cond_dim, facet_args = list(ncol = 5))[[1]] +
  labs(fill = "Speaking\nlevel",  colour = "Speaking\nlevel", x = "Dimension 1 (language attitudes)")

ggsave("img/speak-brm-dim.png", width = 7, height = 4)
```


## Read and write

```{r rw-bm}
rw_bm <- brm(
  read_write ~
    dim_1 * spaces_prop,
  data = emilianto_attitude %>% filter(language == "Emilian"),
  family = bernoulli(),
  # Same priors as comprehend_bm
  prior = comprehend_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/cache/rw_bm"
)
```

```{r rw-bm-sum}
rw_bm
```


```{r rw-bm-cond}
plot(conditional_effects(rw_bm, effects = "spaces_prop:dim_1"))[[1]] +
  labs(fill = "Read-write\nlevel", colour = "Read-write\nlevel", x = "Proportion of spaces")

ggsave("img/rw-brm-spaces.png", width = 7, height = 4)
conditional_effects(rw_bm, effects = "dim_1:spaces_prop")
```

## Age

```{r comprehend-bm-age}
comprehend_bm_age <- brm(
  comprehend ~
    language * age,
  data = emilianto_attitude,
  family = cumulative(),
  # Same priors as comprehend_bm
  prior = comprehend_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/cache/comprehend_bm_age"
)
```

```{r comprehend-cond-age}
conditions <- expand.grid(
  language = c("Emilian", "Esperanto")
)

rownames(conditions) <- unite(conditions, "cond__", everything()) %>% pull(cond__)

ubm_cond_age <- conditional_effects(comprehend_bm_age, effects = "age", categorical = TRUE, conditions = conditions)
plot(ubm_cond_age, facet_args = list(ncol = 5))
```


```{r speak-bm-age}
speak_bm_age <- brm(
  speak ~
    language * age,
  data = emilianto_attitude,
  family = cumulative(),
  # Same priors as comprehend_bm
  prior = comprehend_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/cache/speak_bm_age"
)
```

```{r speak-bm-cond-age}
conditions <- expand.grid(
  language = c("Emilian", "Esperanto")
)

rownames(conditions) <- unite(conditions, "cond__", everything()) %>% pull(cond__)

sbm_cond_age <- conditional_effects(speak_bm_age, effects = "age", categorical = TRUE, conditions = conditions)
plot(sbm_cond_age, facet_args = list(ncol = 5))
```


## Mean competence

```{r meanc}
emilianto_attitude %>%
  ggplot(aes(mean_comp)) +
  geom_density() +
  geom_rug()
```

```{r meanc-jitter}
emilianto_attitude %>%
  ggplot(aes(language, mean_comp)) +
  geom_jitter(width = 0.2, alpha = 0.5, height = 0) +
  expand_limits(y = 0)
```   

```{r mean-jitter-dim}
emilianto_attitude %>%
  ggplot(aes(language, mean_comp, colour = dim_1)) +
  geom_jitter(width = 0.2, alpha = 0.5, height = 0) +
  expand_limits(y = 0)
```

```{r meanc-dim1}
emilianto_attitude %>%
  ggplot(aes(dim_1, mean_comp)) +
  geom_point() +
  facet_grid(~ language)
```

```{r meanc-bm}
mean_comp_bm <- brm(
  mean_comp ~
    language *
    dim_1,
  data = emilianto_attitude,
  family = zero_one_inflated_beta(),
  backend = "cmdstanr",
  cores = 4,
  file = "./data/cache/mean_comp_bm"
)

mean_comp_bm
```

```{r meanc-bm-cond}
conditional_effects(mean_comp_bm, effects = "dim_1:language")
```


## Rural vs urban

### comprehend

```{r emilian}
emilian <- emilianto %>%
  filter(language == "Emilian")
```

```{r em-rur}
emil_rur <- emilian %>%
  mutate(
    ru_ur = ifelse(
      str_detect(birth_place, "-RU"), "rural",
      ifelse(
        str_detect(birth_place, "-UR"), "urban",
        NA
      )
    )
  )

emil_rur_clean <- emil_rur %>%
  dplyr::select(id, comprehend, speak, read_write, educated:familiar, ru_ur) %>%
  mutate(
    comprehend = ordered(comprehend, levels = c("NO", "AL", "50/50", "G", "VG")),
    speak = ordered(speak, levels = c("NO", "AL", "50/50", "G", "VG")),
    across(educated:familiar, ~ as.ordered(.x))
  ) %>%
 drop_na()
```

```{r comprehend-rur}
comprehend_rur <- emil_rur_clean %>%
  count(comprehend, ru_ur) %>%
  group_by(ru_ur) %>%
  mutate(
    ru_ur_prop = n / sum(n)
  ) %>%
  ungroup() %>%
  mutate(
    comprehend = factor(comprehend, ordered = FALSE),
    ru_ur = factor(ru_ur, levels = c("rural", "urban"))
  )
```

```{r comprehend-rur-priors}
get_prior(
  n ~
    comprehend * ru_ur,
  data = comprehend_rur,
  family = Beta()
)
```

```{r comprehend-rur-bm}
comprehend_rur_priors <- c(
  prior(normal(0, 3), class = Intercept),
  prior(normal(0, 1), class = b)
)

comprehend_rur_bm <- brm(
  ru_ur_prop ~
    comprehend * ru_ur,
  data = comprehend_rur,
  family = Beta(),
  prior = comprehend_rur_priors,
  backend = "cmdstanr",
  cores = 4,
  threads = threading(2),
  file = "./data/cache/comprehend_rur_bm"
)

comprehend_rur_bm
```

```{r comprehend-rur-cond}
conditional_effects(comprehend_rur_bm, "comprehend:ru_ur")
```

### Speak

```{r speak-rur}
speak_rur <- emil_rur_clean %>%
  count(speak, ru_ur) %>%
  group_by(ru_ur) %>%
  mutate(
    ru_ur_prop = n / sum(n)
  ) %>%
  ungroup() %>%
  mutate(
    speak = factor(speak, ordered = FALSE),
    ru_ur = factor(ru_ur, levels = c("rural", "urban"))
  )
```


```{r speak-rur-bm}
speak_rur_bm <- brm(
  ru_ur_prop ~
    speak * ru_ur,
  data = speak_rur,
  family = Beta(),
  prior = comprehend_rur_priors,
  backend = "cmdstanr",
  cores = 4,
  file = "./data/cache/speak_rur_bm"
)

speak_rur_bm
```

```{r speak-rur-cond}
conditional_effects(speak_rur_bm, "speak:ru_ur")
```

### Read and write

```{r rw-rur}
rw_rur <- emil_rur_clean %>%
  count(read_write, ru_ur) %>%
  group_by(ru_ur) %>%
  mutate(
    ru_ur_prop = n / sum(n)
  ) %>%
  ungroup() %>%
  mutate(
    ru_ur = factor(ru_ur, levels = c("rural", "urban"))
  )
```


```{r rw-rur-bm}
rw_rur_bm <- brm(
  ru_ur_prop ~
    read_write * ru_ur,
  data = rw_rur,
  family = Beta(),
  prior = comprehend_rur_priors,
  backend = "cmdstanr",
  cores = 4,
  file = "./data/cache/rw_rur_bm"
)

rw_rur_bm
```

```{r rw-rur-cond}
conditional_effects(rw_rur_bm, "read_write:ru_ur")
```

## SEM

```{r}
emilianto_attitude <- emilianto_attitude %>%
  mutate(
    dim_1_z = as.vector(scale(dim_1)),
    spaces_prop_z = as.vector(scale(spaces_prop)),
    speak_ot = speak
  )

contrasts(emilianto_attitude$speak_ot) <- "contr.treatment"
```


```{r eval=FALSE}
bf_1 <- bf(speak ~ language * dim_1_z * spaces_prop_z) + cumulative()
bf_2 <- bf(comprehend ~ language * dim_1_z * spaces_prop_z) + cumulative()
bf_3 <- bf(dim_1_z ~ language * spaces_prop_z * (speak + comprehend)) + gaussian()
bf_4 <- bf(spaces_prop_z ~ language * dim_1_z * (speak + comprehend)) + zero_inflated_beta()

sem_1 <- brm(
  bf_1 + bf_2 + bf_3 + bf_4,
  data = emilianto_attitude,
  cores = 4,
  threads = threading(2),
  backend = "cmdstanr"
)
```


```{r eval=FALSE}
bf_1 <- bf(speak_ot ~ language * dim_1_z * spaces_prop_z) + cumulative()
bf_3 <- bf(dim_1_z ~ language * spaces_prop_z * speak_ot) + gaussian()
bf_4 <- bf(spaces_prop_z ~ language * dim_1_z * speak_ot) + gaussian()

sem_2_prior <- c(
  prior(normal(0, 10), class = Intercept),
  prior(normal(0, 10), class = b),
  prior(cauchy(0, 1), class = sigma, resp = "spacespropz"),
  prior(cauchy(0, 1), class = sigma, resp = "dim1z")
)

sem_2 <- brm(
  bf_1 + bf_3 + bf_4,
  data = emilianto_attitude,
  cores = 4,
  threads = threading(2),
  backend = "cmdstanr",
  file = "data/cache/sem_2",
  prior = sem_2_prior
)
```

```{r eval=FALSE}
summary(sem_2)
```

```{r eval=FALSE}
conditions <- expand.grid(
  language = c("Emilian", "Esperanto")
)

rownames(conditions) <- unite(conditions, "cond__", everything()) %>% pull(cond__)

sbm_cond_sp <- conditional_effects(sem_2, effects = c("speak_ot", "dim_1_z"), resp = "spacespropz", conditions = conditions)
plot(sbm_cond_sp, facet_args = list(ncol = 4))
```
